bin/ghcr is a general purpose code tester
written in Haskell. It takes files that have pure functional code and
adds a `main = do ...` header and writes it to a temp file that it can
call runghc on.

You give it a code file path and an expression string to run.

It has 3 modes of operation:

* Run an expression from a file
* Run an expression from a file that reads stdin and writes stdout
* Run the test function from a file, that has a series of unit tests

See the comments in the code for example usage.